Refactor and more tests for CodeGeneratedQueryStructure
