+++ community/cypher/cypher/src/test/scala/org/neo4j/cypher/internal/compiler/CypherCompilerPerformanceTest.scala
@@ -21,10 +21,10 @@ package org.neo4j.cypher.internal.compiler
 
 import org.neo4j.cypher.GraphDatabaseFunSuite
 import org.neo4j.cypher.internal.CypherCompiler.{CLOCK, DEFAULT_QUERY_PLAN_TTL, DEFAULT_STATISTICS_DIVERGENCE_THRESHOLD}
-import org.neo4j.cypher.internal.compatibility.{WrappedMonitors3_1, WrappedMonitors3_0}
+import org.neo4j.cypher.internal.compatibility.WrappedMonitors3_1
 import org.neo4j.cypher.internal.compiler.v3_1.tracing.rewriters.RewriterStepSequencer
 import org.neo4j.cypher.internal.compiler.v3_1.{CypherCompilerFactory, InfoLogger, _}
-import org.neo4j.cypher.internal.spi.v3_1.GeneratedQueryStructure
+import org.neo4j.cypher.internal.spi.v3_1.codegen.GeneratedQueryStructure
 
 class CypherCompilerPerformanceTest extends GraphDatabaseFunSuite {
