+++ community/cypher/cypher/src/test/scala/org/neo4j/cypher/internal/compiler/v3_1/CypherCompilerAstCacheAcceptanceTest.scala
@@ -27,7 +27,7 @@ import org.neo4j.cypher.internal.compiler.v3_1.executionplan.ExecutionPlan
 import org.neo4j.cypher.internal.compiler.v3_1.tracing.rewriters.RewriterStepSequencer
 import org.neo4j.cypher.internal.frontend.v3_1.ast.Statement
 import org.neo4j.cypher.internal.frontend.v3_1.test_helpers.CypherFunSuite
-import org.neo4j.cypher.internal.spi.v3_1.GeneratedQueryStructure
+import org.neo4j.cypher.internal.spi.v3_1.codegen.GeneratedQueryStructure
 import org.neo4j.graphdb.config.Setting
 import org.neo4j.graphdb.factory.GraphDatabaseSettings
 import org.neo4j.logging.AssertableLogProvider.inLog
